+++ apps/app/src/server/models/workflow.integ.ts
@@ -40,7 +40,7 @@ describe('Workflow', () => {
   });
 
   describe('hasInprogressWorkflowInTargetPage()', () => {
-    test('Should return true if there are in-progress workflows on the page」', async() => {
+    test('Should return true if there are in-progress workflows on the page', async() => {
       // when
       const hasInprogressWorkflowInTargetPage = await Workflow.hasInprogressWorkflowInTargetPage(page1);
