+++ apps/app/src/server/service/workflow.integ.ts
@@ -1,7 +1,7 @@
 import mongoose from 'mongoose';
 
 import {
-  IWorkflowReq, IWorkflowApproverGroupReq, WorkflowStatus, WorkflowApproverStatus, WorkflowApprovalType,
+  IWorkflowReq, WorkflowStatus, WorkflowApproverStatus, WorkflowApprovalType,
 } from '~/interfaces/workflow';
 import Workflow from '~/server/models/workflow';
 
@@ -33,9 +33,10 @@ describe('WorkflowService', () => {
   describe('.createWorkflow', () => {
     test('Should be able to create a workflow', async() => {
 
+      // setup
       const workflow: IWorkflowReq = {
-        creator: creator._id,
-        pageId: page1._id,
+        creator,
+        pageId: page1,
         status: WorkflowStatus.INPROGRESS,
         name: 'page1 workflow',
         comment: 'comment',
@@ -44,11 +45,11 @@ describe('WorkflowService', () => {
             approvalType: WorkflowApprovalType.AND,
             approvers: [
               {
-                user: approver1._id,
+                user: approver1,
                 status:  WorkflowApproverStatus.NONE,
               },
               {
-                user: approver2._id,
+                user: approver2,
                 status: WorkflowApproverStatus.NONE,
               },
             ],
@@ -77,16 +78,16 @@ describe('WorkflowService', () => {
             approvalType: WorkflowApprovalType.AND,
             approvers: [
               {
-                user: approver1._id,
+                user: approver1,
                 status:  WorkflowApproverStatus.NONE,
               },
               {
-                user: approver2._id,
+                user: approver2,
                 status: WorkflowApproverStatus.NONE,
               },
             ],
           },
-        ] as any,
+        ],
       };
 
       // when
@@ -94,93 +95,6 @@ describe('WorkflowService', () => {
 
       // then
       await expect(result).rejects.toThrow('An in-progress workflow already exists');
-    });
-
-
-    describe('.validateApproverGroups', () => {
-      test('Should fail when setting approverGroup.approveType to "OR" when approverGroup.approvers.length is 1', () => {
-
-        // setup
-        const approverGroups: IWorkflowApproverGroupReq[] = [
-          {
-            approvalType: WorkflowApprovalType.OR,
-            approvers: [{ user: creator }],
-          },
-        ];
-
-        // when
-        const caller = () => WorkflowService.validateApproverGroups(true, creator._id, approverGroups);
-
-        // then
-        expect(caller).toThrow('approverGroup.approvalType cannot be set to "OR" when approverGroup.approvers.length is 1');
-      });
-
-      test('Should fail when attempting to set a user as an approver who is already an approver', () => {
-
-        // setup
-        const approverGroups: IWorkflowApproverGroupReq[] = [
-          {
-            approvalType: WorkflowApprovalType.OR,
-            approvers: [
-              {
-                user: approver1,
-              },
-              {
-                user: approver1,
-              },
-            ],
-          },
-        ];
-
-        // when
-        const caller = () => WorkflowService.validateApproverGroups(true, creator._id, approverGroups);
-
-        // then
-        expect(caller).toThrow('Cannot set the same approver within Workflow.ApproverGroups. Also, Workflow.creator cannot be set as an approver.');
-      });
-
-      test('Should fail when attempting to set the workflow creator as an approver', () => {
-
-        // setup
-        const approverGroups: IWorkflowApproverGroupReq[] = [
-          {
-            approvalType: WorkflowApprovalType.AND,
-            approvers: [
-              {
-                user: creator,
-              },
-            ],
-          },
-        ];
-
-        // when
-        const caller = () => WorkflowService.validateApproverGroups(true, creator._id, approverGroups);
-
-        // then
-        expect(caller).toThrow('Cannot set the same approver within Workflow.ApproverGroups. Also, Workflow.creator cannot be set as an approver.');
-      });
-
-      test('Should fail when setting approver.status to anything other than "NONE" during workflow creation', () => {
-
-        // setup
-        const approverGroups: IWorkflowApproverGroupReq[] = [
-          {
-            approvalType: WorkflowApprovalType.AND,
-            approvers: [
-              {
-                user: approver1,
-                status: 'APPROVE',
-              },
-            ],
-          },
-        ];
-
-        // when
-        const caller = () => WorkflowService.validateApproverGroups(true, creator._id, approverGroups);
-
-        // then
-        expect(caller).toThrow('Cannot set approver.status to anything other than "NONE" during creation');
-      });
     });
   });
 });
